refactor(frontend): extract default app shell
